Move alertObserver into internal/search/alert
